refactor(nimbus-mcp): convert build scripts to .ts and integrate icon catalog into prebuild
Convert .mjs scripts to .ts for consistency. Refactor build-icon-catalog to
export an async function and add it as a prebuild step, removing the separate
build command from the build pipeline.
