+++ packages/nimbus-mcp/package.json
@@ -13,10 +13,9 @@
     }
   },
   "scripts": {
-    "prebuild": "node scripts/prebuild.mjs",
-    "build": "pnpm prebuild && tsx scripts/build-icon-catalog.ts && tsup",
+    "prebuild": "tsx scripts/prebuild.ts",
+    "build": "pnpm prebuild && tsup",
     "dev": "tsx src/index.ts",
-    "catalog:icons": "tsx scripts/build-icon-catalog.ts",
     "mcp-inspector": "mcp-inspector tsx src/index.ts",
     "mcp-inspector:cli": "mcp-inspector --cli tsx src/index.ts"
   },
