+++ packages/nimbus-mcp/scripts/copy-docs-data.ts
